# Speed up parsers