--- a/RetailCoder.VBE/Inspections/FunctionReturnValueNotUsedInspection.cs
+++ b/RetailCoder.VBE/Inspections/FunctionReturnValueNotUsedInspection.cs
@@ -25,15 +25,15 @@ namespace Rubberduck.Inspections
         {
             // Note: This inspection does not find dictionary calls (e.g. foo!bar) since we do not know what the
             // default member is of a class.
-            //var interfaceMembers = UserDeclarations.FindInterfaceMembers().ToList();
-            //var interfaceImplementationMembers = UserDeclarations.FindInterfaceImplementationMembers();
-            //var functions = UserDeclarations.Where(function => function.DeclarationType == DeclarationType.Function).ToList();
-            //var interfaceMemberIssues = GetInterfaceMemberIssues(interfaceMembers);
-            //var nonInterfaceFunctions = functions.Except(interfaceMembers.Union(interfaceImplementationMembers));
-            //var nonInterfaceIssues = GetNonInterfaceIssues(nonInterfaceFunctions);
-            //return interfaceMemberIssues.Union(nonInterfaceIssues);
-            // Temporarily disabled until fix for lack of context because of new resolver is found...
-            return new List<InspectionResultBase>();
+            var interfaceMembers = UserDeclarations.FindInterfaceMembers().ToList();
+            var interfaceImplementationMembers = UserDeclarations.FindInterfaceImplementationMembers();
+            var functions = UserDeclarations.Where(function => function.DeclarationType == DeclarationType.Function).ToList();
+            var interfaceMemberIssues = GetInterfaceMemberIssues(interfaceMembers);
+            var nonInterfaceFunctions = functions.Except(interfaceMembers.Union(interfaceImplementationMembers));
+            var nonInterfaceIssues = GetNonInterfaceIssues(nonInterfaceFunctions);
+            return interfaceMemberIssues.Union(nonInterfaceIssues);
+            //// Temporarily disabled until fix for lack of context because of new resolver is found...
+            //return new List<InspectionResultBase>();
         }
 
         private IEnumerable<FunctionReturnValueNotUsedInspectionResult> GetInterfaceMemberIssues(IEnumerable<Declaration> interfaceMembers)
@@ -128,25 +128,40 @@ namespace Rubberduck.Inspections
 
         private bool IsCallStmt(IdentifierReference usage)
         {
-            return usage.Context.Parent is VBAParser.CallStmtContext;
+            var callStmt = ParserRuleContextHelper.GetParent<VBAParser.CallStmtContext>(usage.Context);
+            if (callStmt == null)
+            {
+                return false;
+            }
+            var argumentList = CallStatement.GetArgumentList(callStmt);
+            if (argumentList == null)
+            {
+                return true;
+            }
+            bool isUsedAsArgumentThusReturnValueIsUsed = ParserRuleContextHelper.HasParent(usage.Context, argumentList);
+            return !isUsedAsArgumentThusReturnValueIsUsed;
         }
 
         private bool IsLet(IdentifierReference usage)
         {
-            if (!(usage.Context.Parent is VBAParser.LetStmtContext))
+            var letStmt = ParserRuleContextHelper.GetParent<VBAParser.LetStmtContext>(usage.Context);
+            if (letStmt == null)
             {
                 return false;
             }
-            return ((VBAParser.LetStmtContext)usage.Context.Parent).lExpression() == usage.Context;
+            bool isLetAssignmentTarget = letStmt == usage.Context;
+            return isLetAssignmentTarget;
         }
 
         private bool IsSet(IdentifierReference usage)
         {
-            if (!(usage.Context.Parent is VBAParser.SetStmtContext))
+            var setStmt = ParserRuleContextHelper.GetParent<VBAParser.SetStmtContext>(usage.Context);
+            if (setStmt == null)
             {
                 return false;
             }
-            return ((VBAParser.SetStmtContext)usage.Context.Parent).lExpression() == usage.Context;
+            bool isSetAssignmentTarget = setStmt == usage.Context;
+            return isSetAssignmentTarget;
         }
     }
 }