--- a/Rubberduck.Parsing/Binding/ExpressionClassification.cs
+++ b/Rubberduck.Parsing/Binding/ExpressionClassification.cs
@@ -2,6 +2,7 @@
 {
     public enum ExpressionClassification
     {
+        ResolutionFailed,
         Value,
         Variable,
         Property,