… RetailCoder.VBE/Inspections/RemoveExplicitCallStatmentQuickFix.cs
… RetailCoder.VBE/Inspections/SelfAssignedDeclarationInspection.cs
… RetailCoder.VBE/Inspections/VariableNotAssignedInspection.cs
… RetailCoder.VBE/Inspections/VariableTypeNotDeclaredInspection.cs
… RetailCoder.VBE/Inspections/VariableTypeNotDeclaredInspectionResult.cs
… RetailCoder.VBE/Navigation/CodeExplorer/CodeExplorerMemberViewModel.cs
… RetailCoder.VBE/Refactorings/ExtractMethod/ExtractMethodRefactoring.cs
… RetailCoder.VBE/Refactorings/MoveCloserToUsage/MoveCloserToUsageRefactoring.cs
… RetailCoder.VBE/Refactorings/RemoveParameters/RemoveParametersRefactoring.cs
… RetailCoder.VBE/Refactorings/ReorderParameters/ReorderParametersRefactoring.cs
… RetailCoder.VBE/UI/Command/ShowParserErrorsCommand.cs
… Rubberduck.Parsing/Annotations/AnnotationService.cs
… Rubberduck.Parsing/Binding/BinaryOpDefaultBinding.cs
… Rubberduck.Parsing/Binding/BindingService.cs
… Rubberduck.Parsing/Binding/DefaultBindingContext.cs
… Rubberduck.Parsing/Binding/ExpressionName.cs
… Rubberduck.Parsing/Binding/IBindingContext.cs
… Rubberduck.Parsing/Binding/IndexDefaultBinding.cs
… Rubberduck.Parsing/Binding/MemberAccessDefaultBinding.cs
… Rubberduck.Parsing/Binding/MemberAccessProcedurePointerBinding.cs
… Rubberduck.Parsing/Binding/MemberAccessTypeBinding.cs
… Rubberduck.Parsing/Binding/NewTypeBinding.cs
… Rubberduck.Parsing/Binding/ParenthesizedDefaultBinding.cs
… Rubberduck.Parsing/Binding/ProcedurePointerBindingContext.cs
… Rubberduck.Parsing/Binding/ResolutionFailedExpression.cs
… Rubberduck.Parsing/Binding/SimpleNameDefaultBinding.cs
… Rubberduck.Parsing/Binding/SimpleNameProcedurePointerBinding.cs
… Rubberduck.Parsing/Binding/SimpleNameTypeBinding.cs
… Rubberduck.Parsing/Binding/StatementContext.cs
… Rubberduck.Parsing/Binding/StatementResolutionContext.cs