Changes:
1. Add support for foreign identifiers (#1521). Only the first "level" is considered and simply treated as a normal identifier.
2. Speed up all parsers by rewriting the grammar (#1502). The 65k comments example parses in 400ms here. The parser tests run in 2secs now. The preprocessor tests in 1sec. Modules in smaller test projects are usually parsed in <100ms. If the faster approach fails, it falls back automatically to the slower approach. This way even if there's a combination that requires the slow approach it won't fail. It does get logged however so we can fix it (currently simply using Debug.WriteLine).
3. Integrated expressions into the main grammar. There's no separate expression grammar anymore. Because of that valueStmt and rules like "implicitCallStmt_InStmt" are no more. There's no need (in most cases) to reparse expressions anymore either.

Please review and test with as much code as possible. All tests pass but since quite a few things have changed there might be some subtle things I've missed.