@retailcoder 2016-05-21
Awesome work!! Only missing #1539 with block resolution and we'll be good for a beta I think.