@autoboosh 2016-05-21
With block resolution should actually already work. Can you tell me what exactly doesn't work?

What I forgot to mention is that identifier references are now added even if part of an expression fails.