@autoboosh 2016-05-21
No problem! Please tell me how long the parser/preprocessor tests take on your system. I wonder how long they take on @grleachman's PC. Hopefully under 1 minute...